% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute.sigma.vals}
\alias{compute.sigma.vals}
\title{Find Average Sigma Values Based On Stability Class and Total Distance Traveled}
\usage{
compute.sigma.vals(stab.class, total.dist)
}
\arguments{
\item{stab.class}{A character vector representing the stability class(es) ("A" to "F").}

\item{total.dist}{A numeric value representing the distance traveled in --units--.}
}
\value{
A numeric vector representing the average sigma values over the stability classes passed to the function.
}
\description{
This function calculates the average sigma values based on stability class and the total distance traveled.
             It reports the averages of the sigma values dependent on the inputs.
}
\examples{
compute.sigma.vals(A, 0.7)
}
