% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{faceted_time_series_plot}
\alias{faceted_time_series_plot}
\title{Faceted Time Series Plot of Methane Concentrations and Wind Data}
\usage{
faceted_time_series_plot(
  sensor_concentrations,
  sensor_coords,
  wind_data,
  start_time,
  end_time,
  output_dt
)
}
\arguments{
\item{sensor_concentrations}{Data frame. Output from simulate_sensor_mode().}

\item{sensor_coords}{A data frame or matrix containing sensor locations.}

\item{wind_data}{A list with wind_u and wind_v components.}

\item{start_time}{POSIXct start of simulation.}

\item{end_time}{POSIXct end of simulation.}

\item{output_dt}{Time step (in seconds) for aligning wind data with concentration data.}
}
\value{
A ggplot object: faceted concentration plot + single wind rose plot.
}
\description{
This function creates a faceted bubble plot of methane concentrations and a shared wind rose plot.
}
\examples{
\dontrun{
# Assuming 'sensor_concentrations' is a data frame obtained from simulate_sensor_mode()
output_dt <- 60
start_time <- as.POSIXct("2024-01-01 12:00:00")
end_time <- as.POSIXct("2024-01-01 13:00:00")
wind_data <- list(
  wind_u = runif(3601, min = -3, max = 0.7),
  wind_v = runif(3601, min = -3, max = 1.5)
  )

sensor_coords <- matrix(c(
-6.525403221327715e-15, -35.52264, 2.01775,
-2.15, -33.52264, 2
), ncol = 3, byrow = TRUE)

faceted_time_series_plot(sensor_concentrations, sensor_coords,
wind_data, start_time, end_time, output_dt
)
}
}
