% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuff.R
\name{gaussian_puff}
\alias{gaussian_puff}
\title{Gaussian Puff Concentration Calculation}
\usage{
gaussian_puff(x, y, z, t, q, u, v, z0, wind_speed, sim_time)
}
\arguments{
\item{x}{Numeric. The x-coordinate (east-west) where the concentration is calculated.}

\item{y}{Numeric. The y-coordinate (north-south) where the concentration is calculated.}

\item{z}{Numeric. The z-coordinate (height) where the concentration is calculated.}

\item{t}{Numeric. The time elapsed since the puff was released.}

\item{q}{Numeric. The emission rate of the pollutant.}

\item{u}{Numeric. The wind speed component in the x (east-west) direction.}

\item{v}{Numeric. The wind speed component in the y (north-south) direction.}

\item{z0}{Numeric. The release height of the puff into the atmosphere.}

\item{wind_speed}{Numeric. The total wind speed, calculated as the magnitude of the wind vector.}

\item{sim_time}{POSIXct. The current simulation time, used to determine day or night for stability classification.}
}
\value{
Numeric. The pollutant concentration at the specified (x, y, z) location and time `t`.

Quantities corresponding to concentration at sensor point(s)
}
\description{
Calculates the concentration of an emission event at a specified location and time due to a Gaussian puff.

This function uses wind speed and direction components, advection adjustments, and stability class calculations to
  accurately measure the dispersion of a puff in the atmosphere.
}
\examples{
gaussian_puff(x, y, z, t, q, u, v, z0, wind_speed, sim_time)
}
