% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_stab_class}
\alias{get_stab_class}
\title{Determine Stability Class Based on Wind Speed and Time of Day}
\usage{
get_stab_class(U, time)
}
\arguments{
\item{U}{A numeric value representing the wind speed in meters per second.}

\item{time}{A time value that is used to determine whether it's day or night.}
}
\value{
A character vector representing the stability class(es) ("A" to "F").
}
\description{
This function calculates the stability class based on wind speed (U) and the time of day.
             It categorizes the atmosphere's stability as one of several classes (A-F) depending on the inputs.
}
\examples{
\dontrun{
get_stab_class(3, 12)
}
}
