% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gpuff}
\alias{gpuff}
\title{Gaussian Puff Concentration Calculation}
\usage{
gpuff(Q, stab.class, x.p, y.p, x.r.vec, y.r.vec, z.r.vec, total.dist, H, U)
}
\arguments{
\item{Q}{Numeric. The emission rate of the pollutant.}

\item{stab.class}{Character vector. Represents the stability class(es) ("A" to "F").}

\item{x.p}{Numeric.  Unsure}

\item{y.p}{Numeric. Unusure}

\item{x.r.vec}{Numeric. The x-coordinate (east-west) where the concentration is calculated.}

\item{y.r.vec}{Numeric. The y-coordinate (north-south) where the concentration is calculated.}

\item{z.r.vec}{Numeric. The z-coordinate (height) where the concentration is calculated.}

\item{total.dist}{Numeric. The total distance the puff has traveled in m.}

\item{H}{Numeric. Unsure}

\item{U}{Numeric.  Value representing the wind speed in meters per second.}
}
\value{
Numeric. The pollutant concentration at the specified (x, y, z) location and time `t`.

Quantities corresponding to concentration at sensor point(s)
}
\description{
Calculates the concentration of an emission event at a specified location and time due to a Gaussian puff.

This function uses wind speed and direction components, advection adjustments, and stability class calculations to
  accurately measure the dispersion of a puff in the atmosphere.
}
\examples{
gpuff(Q, stab.class, x.p, y.p, x.r.vec, y.r.vec, z.r.vec, total.dist, H, U)
}
