% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_day}
\alias{is_day}
\title{Determine Whether a Time is During the Day}
\usage{
is_day(time)
}
\arguments{
\item{time}{A time value that is used to determine whether it's day or night.}
}
\value{
A character T or F representing whether or not it is daytime.
}
\description{
This function checks the time and classifies it as day or not
}
\examples{
\dontrun{
is_day(8)
}
}
