% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{wind_vector_convert}
\alias{wind_vector_convert}
\title{Convert between (ws, wd) and (u,v)}
\usage{
wind_vector_convert(wind_speeds,wind_directions)
}
\arguments{
\item{wind_speeds}{A list of float values of wind speeds in m/s at each time stamp}

\item{wind_directions}{A list of float values of wind directions in degrees at each time stamp following
the conventional definition: 0 -> wind blowing from North, 90 -> E, 180 -> S, 270 -> W}
}
\value{
Quantities corresponding to the conversion direction
}
\description{
This function converts between coordinate systems by changing from degrees to radians
}
\examples{
\dontrun{
wind_vector_convert(speed_vec,direction_vec)
}
}
