% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_sigma_vals}
\alias{compute_sigma_vals}
\title{Compute Sigma Values Based on Stability Class and Distance}
\usage{
compute_sigma_vals(stab_class, total_dist)
}
\arguments{
\item{stab_class}{Character vector of stability classes ("A" to "F")}

\item{total_dist}{Numeric vector of distances in km (must match length of stab_class or be scalar)}
}
\value{
2-row matrix with sigma_y (row 1) and sigma_z (row 2) values
}
\description{
Compute Sigma Values Based on Stability Class and Distance
}
\examples{
\dontrun{
out <- compute_sigma_vals(A, 0.7)
}
}
