% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{faceted_time_series_plot}
\alias{faceted_time_series_plot}
\title{Faceted Time Series Plot of Methane Concentrations and Wind Data}
\usage{
faceted_time_series_plot(sensor_concentrations, wind_data, time_sequence)
}
\arguments{
\item{sensor_concentrations}{Data frame. Output from a sensor simulation function,
which must include a column named "Group.1" which contains the timestamps (e.g., "YYYY-MM-DD HH:MM:SS") and a column "Sensor_1"
for the sensor concentration values.}

\item{wind_data}{A list containing wind data with components u and v}

\item{time_sequence}{A POSIXct vector of time steps corresponding to the wind data.}
}
\value{
A ggplot object with faceted time series plots of methane concentrations and wind data.
}
\description{
This function creates a faceted plot showing methane concentrations along with wind data over time.
}
\examples{
\dontrun{
# Assuming 'sensor_concentrations' is a data frame obtained from simulate_sensor_mode

wind_data <- list(
  wind_u = runif(length(time_sequence), min = -5, max = 5),
  wind_v = runif(length(time_sequence), min = -5, max = 5)
)

faceted_time_series_plot(sensor_concentrations, wind_data, time_sequence)
}
}
