% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{faceted_time_series_plot2}
\alias{faceted_time_series_plot2}
\title{Alternate version with wind rose at each time step + scatter plot of methane concentration time series}
\usage{
faceted_time_series_plot2(
  sensor_concentrations,
  sensor_coords,
  wind_data,
  start_time,
  end_time,
  output_dt
)
}
\arguments{
\item{sensor_concentrations}{Data frame. Output from a sensor simulation function,
which must include a column named "Group.1" which contains the timestamps (e.g., "YYYY-MM-DD HH:MM:SS") and a column "Sensor_1"
for the sensor concentration values.}

\item{wind_data}{A list containing wind data with components u and v}

\item{start_time}{POSIXct. Start time of the simulation.}

\item{end_time}{POSIXct. End time of the simulation.}

\item{output_dt}{Integer. Desired time resolution (in seconds) for the final output of concentrations.}
}
\value{
A ggplot object with faceted time series plots of methane concentrations and wind rose data.
}
\description{
Alternate version with wind rose at each time step + scatter plot of methane concentration time series
}
