% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gpuff}
\alias{gpuff}
\title{Gaussian Puff Concentration Calculation}
\usage{
gpuff(Q, stab.class, x.p, y.p, x.r.vec, y.r.vec, z.r.vec, total.dist, H, U)
}
\arguments{
\item{Q}{Numeric. Mass per puff. E.g., for 100 puffs/hour of a 100 kg/hr emission, put 1 kg of mass into each puff.}

\item{stab.class}{Character vector. Stability class ("A" to "F").}

\item{x.p}{Numeric. Puff position in the X direction.}

\item{y.p}{Numeric. Puff position in the Y direction.}

\item{x.r.vec}{Numeric. The x-coordinate (east-west) where the concentration is calculated.}

\item{y.r.vec}{Numeric. The y-coordinate (north-south) where the concentration is calculated.}

\item{z.r.vec}{Numeric. The z-coordinate (height) where the concentration is calculated.}

\item{total.dist}{Numeric. The total distance the puff has traveled from the source in m.}

\item{H}{Numeric. Source height.}

\item{U}{Numeric. Wind speed in m/s.}
}
\value{
Numeric. Pollutant concentration at the specified (x, y, z) locations and time `t`.
}
\description{
Calculates the concentration of an emission event at a specified location and time due to a Gaussian puff.

This function uses wind speed and direction components, advection adjustments, and stability class calculations to
  accurately measure the dispersion of a puff in the atmosphere.
}
\examples{
gpuff(Q, stab.class, x.p, y.p, x.r.vec, y.r.vec, z.r.vec, total.dist, H, U)
}
