% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_2d_animated}
\alias{plot_2d_animated}
\title{Plot a 2D Animated Heatmap for Concentration Over Time}
\usage{
plot_2d_animated(
  data,
  grid_coords,
  start,
  end,
  output_dt,
  frames = 100,
  transition = 99,
  save = FALSE,
  interpolate_grid = FALSE,
  granularity = 100
)
}
\arguments{
\item{data}{A matrix or array of grid concentration results from `simulate_grid_mode()`.}

\item{grid_coords}{A list containing the same grid coordinates passed to `simulate_grid_mode()`.}

\item{start}{A character string specifying the start time of the simulation (e.g., "YYYY-MM-DD HH:MM:SS").}

\item{end}{A character string specifying the end time of the simulation (e.g., "YYYY-MM-DD HH:MM:SS").}

\item{output_dt}{A character string or numeric value specifying the time interval between outputs.}

\item{frames}{Numeric. Duration between frames in the animation (milliseconds).}

\item{transition}{Numeric. Duration for transitioning between frames (milliseconds).}

\item{save}{Logical. If `TRUE`, saves the plot as an HTML file named `2D_heatmap.html` and specifies saved location. Default set to `FALSE`.}

\item{interpolate_grid}{Logical. If `TRUE`, applies interpolation to refine grid resolution and make the heatmap smoother. Default `FALSE`.}

\item{granularity}{Numeric. Sets the number of points in the finer grid resolution when `interpolate_grid = TRUE`. Default `100`.}
}
\value{
A `plotly` object representing the animated heatmap.
}
\description{
This function generates a 2D animated heatmap using `plotly` to visualize the
movement of a plume over time. The animation is based on grid concentration data
from `simulate_grid_mode()` output.
}
\examples{
\dontrun{
grid_concentrations <- array(...)  # 3D concentration array
grid_coords <- list(x = ..., y = ..., z = ...)

plot_2d_animated(data = grid_concentrations,
  grid_coords = grid_coords,
  start = "2025-01-01 00:00:00",
  end = "2025-01-01 01:00:00",
  output_dt = "1 min"
  )
}
}
