% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{time_series_plot}
\alias{time_series_plot}
\title{Plot Time Series of Sensor Concentrations}
\usage{
time_series_plot(sensor_concentrations)
}
\arguments{
\item{sensor_concentrations}{A data frame (or matrix) containing the output from
the sensor simulation function (e.g., `simulate_sensor_mode()`). It must include:
\describe{
  \item{Group.1}{A character vector of timestamps in the format `"YYYY-MM-DD HH:MM:SS"`.}
  \item{Sensor_1}{A numeric vector of sensor concentration values corresponding to each timestamp.}
}}
}
\value{
A ggplot object showing the time series of sensor concentrations.
}
\description{
This function plots the time series of sensor concentrations.
}
\examples{
\dontrun{
# Assuming 'sensor_concentrations' is a data frame obtained from simulate_sensor_mode
time_series_plot(sensor_concentrations)
}
}
